fix: move bonus inside modules
