+++ src/modules/bonus/components/BonusPriceTag.tsx
